+++ tbx/docmaker/+docmaker/getDefaultFigureSize.m
@@ -1,0 +1,7 @@
+function s = getDefaultFigureSize()
+%getDefaultFigureSize  Default figure size
+
+p = get( 0, "DefaultFigurePosition" ); % [x y w h]
+s = p(3:4); % [w h]
+
+end % getDefaultFigureSize